+++ src/components/modals/TokenSelectModal.svelte
@@ -1,5 +1,5 @@
 <script>
-  import images from "../config/images.json";
+  import images from "../../config/images.json";
 
   export let callback;
   export let tokens;
